Signout Api & Admin Dashboard

+++ swadmall-admin-app/src/actions/auth.actions.js
@@ -79,6 +79,7 @@ export const login = (user) => {
 
 export const isUserLoggedIn = () => {
     return async dispatch => {
+        const res = await axios.post('/admin/signout')
         const token = localStorage.getItem('token');
         if(token){
             const user = JSON.parse(localStorage.getItem('user'));
@@ -104,11 +105,22 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
+        dispatch({ type: authConstants.LOGOUT_REQUEST});
+        const res = await axios.post('/admin/signout');
+        if(res.status === 200){
+            localStorage.clear();
+            dispatch({
+                type: authConstants.LOGOUT_SUCCESS});
+        }else{
+            dispatch({
+                type: authConstants.LOGIN_FAILURE,
+                payload: { error: res.data.error}
 
-        });
+            
+            })
+
+        }
+
 
     }
